+++ app/view/lib/bolt/start.js
@@ -1,10 +1,33 @@
+/**
+ * Extend String class with a placeholder replacement function
+ *
+ * Placeholder strings must be surrounded by %, start with an uppercase character, followed by on ore more
+ * uppercase characters, numbers or _
+ *
+ * @param {Array} replacements
+ * @returns {String}
+ */
+String.prototype.subst = function(replacements) {
+    return this.replace(/%[A-Z][A-Z0-9_]+%/g, function (placeholder) {
+        return replacements[placeholder] || placeholder;
+    });
+};
 
+/**
+ * Starting point
+ *
+ * @param {Object} $
+ * @returns {undefined}
+ */
 jQuery(function ($) {
     // Get configuration
     var config = $('script[data-config]').first().data('config');
     for (var key in config) {
         bolt[key] = config[key];
     }
+
+    // Get passed in data from Twig function data()
+    bolt.data = $('script[data-jsdata]').first().data('jsdata');
 
     // Initialize objects
     bolt.files = new Files();