+++ src/Application.php
@@ -25,7 +25,7 @@ class Application extends Silex\Application
 
     public function __construct(array $values = array())
     {
-        $values['bolt_version'] = '2.0.2';
+        $values['bolt_version'] = '2.0.3';
         $values['bolt_name'] = '';
 
         parent::__construct($values);
@@ -54,6 +54,9 @@ class Application extends Silex\Application
         // Initialize the 'editlink' and 'edittitle'..
         $this['editlink'] = '';
         $this['edittitle'] = '';
+        
+        // Initialise the JavaScipt data gateway
+        $this['jsdata'] = array();
     }
 
     /**
@@ -121,16 +124,33 @@ class Application extends Silex\Application
      */
     public function initDatabase()
     {
-        $dboptions = $this['config']->getDBOptions();
-
         $this->register(
             new Silex\Provider\DoctrineServiceProvider(),
             array(
-                'db.options' => $dboptions
+                'db.options' => $this['config']->getDBOptions()
             )
         );
 
-        // Do a dummy query, to check for a proper connection to the database.
+        $this->checkDatabaseConnection();
+
+        $this->tweakDatabaseDefaults();
+
+        $this->register(
+            new Silex\Provider\HttpCacheServiceProvider(),
+            array(
+                'http_cache.cache_dir' => $this['resources']->getPath('cache'),
+            )
+        );
+    }
+
+    /**
+     * Do a dummy query, to check for a proper connection to the database.
+     * @throws LowlevelException
+     */
+    protected function checkDatabaseConnection()
+    {
+        $dboptions = $this['db.options'];
+
         try {
             $this['db']->query("SELECT 1;");
         } catch (\PDOException $e) {
@@ -142,6 +162,11 @@ class Application extends Silex\Application
             }
             throw new LowlevelException($error);
         }
+    }
+
+    protected function tweakDatabaseDefaults()
+    {
+        $dboptions = $this['db.options'];
 
         if ($dboptions['driver'] == 'pdo_sqlite') {
             $this['db']->query('PRAGMA synchronous = OFF');
@@ -156,13 +181,6 @@ class Application extends Silex\Application
             $this['db']->query("SET CHARACTER_SET_CONNECTION = 'utf8';");
             $this['db']->query("SET CHARACTER SET utf8;");
         }
-
-        $this->register(
-            new Silex\Provider\HttpCacheServiceProvider(),
-            array(
-                'http_cache.cache_dir' => $this['resources']->getPath('cache'),
-            )
-        );
     }
 
     public function initRendering()
@@ -186,6 +204,18 @@ class Application extends Silex\Application
                 )
             )
         );
+        // Add the Bolt Twig Extension.
+        $this['twig'] = $this->share(
+            $this->extend(
+                'twig',
+                function(\Twig_Environment $twig, $app) {
+                    $twig->addExtension(new TwigExtension($app));
+                    return $twig;
+                }
+            )
+        );
+
+        $this->register(new Provider\SafeTwigServiceProvider());
 
         $this->register(new Provider\RenderServiceProvider());
         $this->register(new Provider\RenderServiceProvider(true));
@@ -257,7 +287,6 @@ class Application extends Silex\Application
             ->register(new Provider\OmnisearchServiceProvider())
             ->register(new Provider\TemplateChooserServiceProvider())
             ->register(new Provider\CronServiceProvider())
-            ->register(new Provider\SafeTwigServiceProvider())
             ->register(new Provider\FilePermissionsServiceProvider())
             ->register(new Controllers\Upload())
             ->register(new Controllers\Extend())
@@ -266,17 +295,9 @@ class Application extends Silex\Application
 
         $this['paths'] = $this['resources']->getPaths();
 
-        $this['twig']->addGlobal('paths', $this['paths']);
-
         // For some obscure reason, and under suspicious circumstances $app['locale'] might become 'null'.
         // Re-set it here, just to be sure. See https://github.com/bolt/bolt/issues/1405
         $this['locale'] = $currentlocale;
-
-        // Add the Bolt Twig functions, filters and tags.
-        $this['twig']->addExtension(new TwigExtension($this));
-        $this['safe_twig']->addExtension(new TwigExtension($this, true));
-
-        $this['twig']->addTokenParser(new SetcontentTokenParser());
 
         // Initialize stopwatch even if debug is not enabled.
         $this['stopwatch'] = $this->share(
@@ -335,38 +356,6 @@ class Application extends Silex\Application
         $this->mount('', new Controllers\Routing());
     }
 
-
-    /**
-     * Add all the global twig variables, like 'user' and 'theme'
-     */
-    private function addTwigGlobals()
-    {
-        $this['twig']->addGlobal('bolt_name', $this['bolt_name']);
-        $this['twig']->addGlobal('bolt_version', $this['bolt_version']);
-
-        $this['twig']->addGlobal('frontend', false);
-        $this['twig']->addGlobal('backend', false);
-        $this['twig']->addGlobal('async', false);
-        $this['twig']->addGlobal($this['config']->getWhichEnd(), true);
-
-        $this['twig']->addGlobal('user', $this['users']->getCurrentUser());
-        $this['twig']->addGlobal('users', $this['users']->getUsers());
-        $this['twig']->addGlobal('config', $this['config']);
-        $this['twig']->addGlobal('theme', $this['config']->get('theme'));
-
-        $this['safe_twig']->addGlobal('bolt_name', $this['bolt_name']);
-        $this['safe_twig']->addGlobal('bolt_version', $this['bolt_version']);
-
-        $this['safe_twig']->addGlobal('frontend', false);
-        $this['safe_twig']->addGlobal('backend', false);
-        $this['safe_twig']->addGlobal('async', false);
-        $this['safe_twig']->addGlobal($this['config']->getWhichEnd(), true);
-
-        $this['safe_twig']->addGlobal('user', $this['users']->getCurrentUser());
-        $this['safe_twig']->addGlobal('theme', $this['config']->get('theme'));
-    }
-
-
     /**
      * Initializes the Console Application that is responsible for CLI interactions.
      */
@@ -387,9 +376,6 @@ class Application extends Silex\Application
     {
         // Start the 'stopwatch' for the profiler.
         $this['stopwatch']->start('bolt.app.before');
-
-        // Set the twig Globals, like 'user' and 'theme'.
-        $this->addTwigGlobals();
 
         if ($response = $this['render']->fetchCachedRequest()) {
             // Stop the 'stopwatch' for the profiler.
@@ -445,12 +431,20 @@ class Application extends Silex\Application
             // Register the toolbar item for the Twig toolbar item.
             $this->register(new Provider\TwigProfilerServiceProvider());
 
-            $this['twig.loader.filesystem']->addPath(
-                $this['resources']->getPath('root') . '/vendor/symfony/web-profiler-bundle/Symfony/Bundle/WebProfilerBundle/Resources/views',
-                'WebProfiler'
-            );
+            $this['twig.loader.filesystem'] = $this->share(
+                $this->extend(
+                    'twig.loader.filesystem',
+                    function(\Twig_Loader_Filesystem $filesystem, $app) {
+                        $filesystem->addPath(
+                            $app['resources']->getPath('root') . '/vendor/symfony/web-profiler-bundle/Symfony/Bundle/WebProfilerBundle/Resources/views',
+                            'WebProfiler'
+                        );
+                        $filesystem->addPath($app['resources']->getPath('app') . '/view', 'BoltProfiler');
 
-            $this['twig.loader.filesystem']->addPath($this['resources']->getPath('app') . '/view', 'BoltProfiler');
+                        return $filesystem;
+                    }
+                )
+            );
 
             // PHP 5.3 does not allow 'use ($this)' in closures.
             $app = $this;
@@ -549,9 +543,6 @@ class Application extends Silex\Application
         $end = $this['config']->getWhichEnd();
 
         $trace = $exception->getTrace();
-
-        // Set the twig Globals, like 'user' and 'theme'.
-        $this->addTwigGlobals();
 
         foreach ($trace as $key => $value) {
             if (!empty($value['file']) && strpos($value['file'], '/vendor/') > 0) {